avatar_fun/requirements.txt
numpy
Pillow
moviepy